race update
