+++ src/app/modules/game/game-routing.module.ts
@@ -1,12 +1,20 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from './home/home.component';
+import { RaceComponent } from './race/race.component';
 
 const routes: Routes = [
   {
     path: 'home',
     component: HomeComponent,
     data:{
+      requiresLogin: true
+    }
+  },
+  {
+    path: 'race', 
+    component: RaceComponent, 
+    data :{
       requiresLogin: true
     }
   }
