+++ src/app/modules/game/race/race.component.scss
@@ -1,5 +1,4 @@
 .speedway{
-    margin-top: -1%;
     background: url("../../../../assets/img/pista-carrera.png") no-repeat 0 0;
     background-repeat: no-repeat;
     background-size: 100% 95%;
